A possible CA
Currently not working properly but could be an alternative. It is pretty much fully threaded. A problem I have run into is a huge increase in display_update processing time (not sure why) and over time performance drops, but can hopefully fix this. By default it will multiplace. To stop this set break delay to about 50 and place delay to around 160.